mars-rover-photos-api/requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-decouple==3.8
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.6
Werkzeug==3.0.0
